A little overhaul of the verification code - started from search, will need a rebase to properly split the commits
